pyproject.toml
[project]
name = "trendradar-mcp"
version = "1.1.0"
description = "TrendRadar MCP Server - 新闻热点聚合工具"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.5,<3.0.0",
    "pytz>=2025.2,<2026.0",
    "PyYAML>=6.0.3,<7.0.0",
    "fastmcp>=2.12.0,<2.14.0",
    "websockets>=13.0,<14.0",
    "boto3>=1.35.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.30.0,<1.0.0",
    "jinja2>=3.1.4,<4.0.0",
    "ruamel.yaml>=0.18.6,<0.19.0",
]

[project.scripts]
trendradar = "mcp_server.server:run_server"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]
